pyproject.toml
[project]
name = "home-bookkeeping"
version = "0.1.0"
description = ""
authors = [{ name = "SkFrozen", email = "frozentw91@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "alembic-postgresql-enum (>=1.7.0,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
